+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * [#2948](https://github.com/ruby-grape/grape/pull/2948): Speed up path-versioned requests and error responses - [@ericproulx](https://github.com/ericproulx).
 * [#2949](https://github.com/ruby-grape/grape/pull/2949): Present an error response with the entity a `desc` `failure:` keyword names - [@ericproulx](https://github.com/ericproulx).
 * [#2947](https://github.com/ruby-grape/grape/pull/2947): Speed up `declared` filling in a missing Hash param - [@ericproulx](https://github.com/ericproulx).
+* [#2950](https://github.com/ruby-grape/grape/pull/2950): Read a configuration key nested in a Hash or an Array in a remountable API instead of raising `NoMethodError` - [@ericproulx](https://github.com/ericproulx).
 * Your contribution here.
 
 ### 4.0.1 (2026-09-15)